refactor(core): move runs module from pitboss-cli (#484 first cut)
Run discovery (RunStatus, RunEntry, runs_base_dir,
resolve_run_dir_by_prefix, collect_run_entries, control_socket_is_live,
format_mtime) is a pure-on-disk read-side concern with no dispatcher
dependencies. Pre-fix it lived in pitboss-cli, which forced pitboss-tui
and pitboss-web to pull in the full dispatcher crate (MCP server,
process spawner, hierarchical orchestration, ...) just to access the
classifier.

Move runs.rs into pitboss-core. pitboss-cli/src/runs.rs is now a thin
`pub use pitboss_core::runs::*;` shim so existing `crate::runs::...`
imports keep working byte-for-byte across the dispatcher's internal
callers (prune, diff, list, analyze, attach, events, audit, main).
pitboss-tui's runs.rs shim and pitboss-web's runs_index.rs switch to
import from pitboss_core::runs directly.

This does NOT remove the pitboss-cli dep from TUI or web — both still
need ~8 other module surfaces (control::protocol, capability_matrix,
mcp::policy, dispatch::events, manifest::resolve, audit, stream,
shared_store). Closing #484 fully requires moving those too, but a
chunk of them carry MCP/approval types that pitboss-core/CLAUDE.md
explicitly forbids hosting (typed ControlEvent, ApprovalRule). The
remainder will need follow-up scoping. This PR is the first move; it
eliminates the heaviest single dependency surface (9 type/function
names across ~50+ call sites).

Adds two regular deps to pitboss-core (`anyhow`, `dirs`) and one dev
dep (`serial_test` for the existing `#[serial(env)]` test). All five
were already workspace deps.

Fixed five clippy::pedantic complaints in the moved code that pitboss-cli
had not enforced: variables-in-format-string, item-after-statements,
needless-Option-wrap, map-then-unwrap-or, stable-sort-primitive.

Refs #484
